+++ src/debug/objects.rs
@@ -9,14 +9,14 @@ use bevy::{color::palettes::css::RED, prelude::*};
 ///
 /// Contains some settings to customize how the `arrow_2d` [Gizmos] will appear.
 #[derive(Resource, Clone)]
-pub struct TiledMapGizmosConfig {
+pub struct TiledDebugObjectsConfig {
     /// Color of the `arrow_2d` [Gizmos]
     pub color: Color,
     /// Length of the `arrow_2d` [Gizmos]
     pub arrow_length: Vec2,
 }
 
-impl Default for TiledMapGizmosConfig {
+impl Default for TiledDebugObjectsConfig {
     fn default() -> Self {
         Self {
             color: bevy::prelude::Color::Srgba(RED),
@@ -33,27 +33,23 @@ impl Default for TiledMapGizmosConfig {
 /// use bevy_ecs_tiled::prelude::*;
 ///
 /// App::new()
-///     .add_plugins(TiledMapDebugPlugin::default());
+///     .add_plugins(TiledDebugObjectsPlugin::default());
 /// ```
 ///
 /// This will display an `arrow_2d` [Gizmos] where your objects are.
 ///
 #[derive(Default, Clone)]
-pub struct TiledMapDebugPlugin {
-    /// Debug gizmos configuration
-    pub gizmos_config: TiledMapGizmosConfig,
-}
-
-impl Plugin for TiledMapDebugPlugin {
+pub struct TiledDebugObjectsPlugin(pub TiledDebugObjectsConfig);
+impl Plugin for TiledDebugObjectsPlugin {
     fn build(&self, app: &mut bevy::prelude::App) {
-        app.insert_resource(self.gizmos_config.clone())
+        app.insert_resource(self.0.clone())
             .add_systems(Update, draw_debug_arrow);
     }
 }
 
 fn draw_debug_arrow(
     q_objects: Query<&GlobalTransform, With<TiledMapObject>>,
-    config: Res<TiledMapGizmosConfig>,
+    config: Res<TiledDebugObjectsConfig>,
     mut gizmos: Gizmos,
 ) {
     for transform in q_objects.iter() {